+++ builder/rollup/rollup.html.mjs
@@ -5,17 +5,19 @@ import jsyaml from 'js-yaml';
 import commonjs from '@rollup/plugin-commonjs';
 import alias from '@rollup/plugin-alias';
 import json from '@rollup/plugin-json';
-import yaml from '@rollup/plugin-yaml';
+// import yaml from '@rollup/plugin-yaml';
 import resolve from '@rollup/plugin-node-resolve';
 // import typescript from '@rollup/plugin-typescript';
 import sucrase from '@rollup/plugin-sucrase';
 
+import dynamicImportVars from '@rollup/plugin-dynamic-import-vars';
 import postcss from 'rollup-plugin-postcss';
-import {terser} from 'rollup-plugin-terser';
+import terser from '@rollup/plugin-terser';
+// import {terser} from 'rollup-plugin-terser';
 
 import nodePolyfills from 'rollup-plugin-polyfill-node';
-import dynamicImportVars from '@rollup/plugin-dynamic-import-vars';
 
+import yaml from '../../tasks/plugin/rollup-plugin-yaml.mjs';
 
 let drop_console = false;
 if(process.env.DROP_CONSOLE === 'true') {