+++ builder/rollup/rollup.nodejs.mjs
@@ -6,12 +6,15 @@ import jsyaml from 'js-yaml';
 
 import alias from '@rollup/plugin-alias';
 import json from '@rollup/plugin-json';
-import yaml from '@rollup/plugin-yaml';
+// import yaml from '@rollup/plugin-yaml';
 // import typescript from '@rollup/plugin-typescript';
 import sucrase from '@rollup/plugin-sucrase';
 import resolve from '@rollup/plugin-node-resolve';
-
 import dynamicImportVars from '@rollup/plugin-dynamic-import-vars';
+
+
+import yaml from '../../tasks/plugin/rollup-plugin-yaml.mjs';
+
 
 let filename;