+++ builder/rollup/rollup.resource.mjs
@@ -6,15 +6,19 @@ import atlas from '../../tasks/plugin/rollup-plugin-atlas.mjs';
 
 import alias from '@rollup/plugin-alias';
 import json from '@rollup/plugin-json';
-import yaml from '@rollup/plugin-yaml';
+// import yaml from '@rollup/plugin-yaml';
 import resolve from '@rollup/plugin-node-resolve';
 
 // import typescript from '@rollup/plugin-typescript';
 // import sucrase from '@rollup/plugin-sucrase';
 import virtual from '@rollup/plugin-virtual';
-import {terser} from 'rollup-plugin-terser';
+import terser from '@rollup/plugin-terser';
+// import {terser} from 'rollup-plugin-terser';
+
 import globImport from 'rollup-plugin-glob-import';
 import dynamicImportVars from '@rollup/plugin-dynamic-import-vars';
+
+import yaml from '../../tasks/plugin/rollup-plugin-yaml.mjs';
 
 
 let templateVersion = 'game.4.0';